[Course exclusions] Type API
The modified data is not used in the frontend so we can simply return a status code
