--- services/frontend/src/redux/courseExclusions.js
@@ -1,23 +1,0 @@
-import { RTKApi } from '@/apiConnection'
-
-const courseExclusionsApi = RTKApi.injectEndpoints({
-  endpoints: builder => ({
-    setCourseExclusion: builder.mutation({
-      query: ({ programmeCode, courseCodes, curriculumVersion }) => ({
-        url: `/v3/course_exclusions/${programmeCode}`,
-        method: 'POST',
-        body: { courseCodes, curriculumVersion },
-      }),
-    }),
-    removeCourseExclusion: builder.mutation({
-      query: ({ programmeCode, courseCodes, curriculumVersion }) => ({
-        url: `/v3/course_exclusions/${programmeCode}`,
-        method: 'DELETE',
-        body: { courseCodes, curriculumVersion },
-      }),
-    }),
-    overrideExisting: false,
-  }),
-})
-
-export const { useSetCourseExclusionMutation, useRemoveCourseExclusionMutation } = courseExclusionsApi
